# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User CRUD API with Authentication",
    "description": "API for managing users, authentication, and projects with JWT-based security",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://test-user-crud-api-215742463493.asia-northeast1.run.app",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "example": "Website Redesign"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500,
            "example": "Complete redesign of the company website"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed"],
            "default": "pending",
            "example": "in_progress"
          },
          "userId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "minItems": 1,
            "example": [1, 2, 3],
            "description": "Array of user IDs assigned to this project (no duplicates allowed)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-27T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-27T10:30:00Z"
          }
        },
        "required": ["id", "title", "status", "userId"]
      },
      "CreateProjectRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[a-zA-Z0-9.\\-'\\s]+$",
            "example": "Website Redesign",
            "description": "Project title"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "example": "Complete redesign of the company website"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed"],
            "default": "pending",
            "example": "pending"
          },
          "userId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "minItems": 1,
            "uniqueItems": true,
            "example": [1, 2],
            "description": "Array of user IDs to assign to this project (must contain at least one user, no duplicates)"
          }
        },
        "required": ["title", "userId"]
      },
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[a-zA-Z0-9.\\-'\\s]+$",
            "example": "Website Redesign"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "example": "Complete redesign of the company website"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed"],
            "example": "in_progress"
          },
          "userId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "minItems": 1,
            "uniqueItems": true,
            "example": [1, 2, 3],
            "description": "Array of user IDs to assign to this project (must contain at least one user, no duplicates)"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error description"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Validation error 1", "Validation error 2"]
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get all projects",
        "description": "Returns a list of all projects (currently unprotected)",
        "operationId": "findAllProjects",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "description": "Add a new project to the database (requires authentication). All user IDs in the userId array must exist.",
        "operationId": "createProject",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "One or more users not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project by ID",
        "description": "Returns a single project by ID (currently unprotected)",
        "operationId": "findProjectById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": ["Projects"],
        "summary": "Update an existing project",
        "description": "Update a project by ID (requires authentication). All user IDs in the userId array must exist.",
        "operationId": "updateProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or ID parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Project not found or one or more users not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete a project",
        "description": "Delete a project by ID (requires authentication)",
        "operationId": "deleteProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted successfully (No Content)"
          },
          "400": {
            "description": "Invalid ID parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}